requirements.txt
tensorflow-cpu==2.13.1
Pillow
numpy